aboutus.html .mainHeading and .topBar
aboutus.html .mainHeading and .topBar are at a good point so figured that I should save. The .mainHeading and acompanying text has been styled and should carry over to the similar pages. Also, I have continued the styling of the .topBar and it now matches the proto for the most part.