Post Code Review Improvements
Kristin helped make my code a bit more bulletproof in following the requirements from the project. For example,  .getBooksByAuthor had to be modified to return distinct authors as it was just returning all authors with duplicates. Also, fixed the .addBooks and deleted the return null as it wasn't required in the project guidlines. I can now work on the bonus assignment and hopefully add UI to make it easier to add books.