Project styling modified, styling has changed still planning stage
I was thinking whether I should use a modal for the user interface that .addsBook and .addsBooks, I am prefering a hidden card w/ a form that expands and hopefullys simulates a modal w/out mobile disruption. I want to show first just a column of basic inputs in a card. with .addBooks