still need local storage and to display funcionality, and architecture
I was checking my f(x)'s and listeners, they are all working as observed in the console. I need to update the table without inputing all of the books already in the library.
Also, I need to implement the required init, bind, and handlers architecture.
Finally, connect the local storage function that I have.