structure in place but commented out for basic testing
functionality was working in the console on a different page but when pasted into this page the console states that $.ajax is not a function. I paste the same f(x) to codepen and it works! There is no difference and it works elsewhere but not on our page, I have checked and jQuery is installed etc. What a mystery!