almost done,  code was correct but error w jQuery CDN
there was some sort of error w the jQuery on the CDN I was using earlier today. It was driving me crazy, my code just seemed correct. Debugger kept saying that the ajax wasn't a function but worked in the console in other pages. I tried everyting, found the bug on line 2 of the cdn.
-able to print requirements to page into cards! almost done progressed far into project, yay